Move project card to components
